Split the tab thunks off the tab reducer
