+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -10,4 +10,5 @@ export * from "./sharing";
 export * from "./ui";
 export * from "./actions";
 export * from "./tabs";
+export * from "./tabs-thunks";
 export * from "./trash";
